+++ Source/Util/DataStream.h
@@ -16,22 +16,22 @@
 typedef struct stream
 {
     uint8_t* data;
-    uint32_t length;
-    uint32_t pos;
+    size_t length;
+    size_t pos;
 } stream_t;
 
-void stream_create(stream_t* stream, uint32_t length);
+void stream_create(stream_t* stream, size_t length);
 void stream_free(stream_t* stream);
-uint32_t stream_left(const stream_t* stream);
-void stream_skip(stream_t* stream, uint32_t skip);
+size_t stream_left(const stream_t* stream);
+void stream_skip(stream_t* stream, size_t skip);
 uint8_t stream_read_u8(stream_t* stream);
 uint16_t stream_read_u16(stream_t* stream);
 uint32_t stream_read_u32(stream_t* stream);
 float stream_read_f(stream_t* stream);
-void stream_read_array(stream_t* stream, void* output, uint32_t length);
-size_t stream_read_string(stream_t* stream, char* output, uint32_t max_length);
+void stream_read_array(stream_t* stream, void* output, size_t length);
+size_t stream_read_string(stream_t* stream, char* output, size_t max_length);
 void stream_write_u8(stream_t* stream, uint8_t value);
 void stream_write_u16(stream_t* stream, uint16_t value);
 void stream_write_u32(stream_t* stream, uint32_t value);
 void stream_write_f(stream_t* stream, float value);
-void stream_write_array(stream_t* stream, const void* array, uint32_t length);
+void stream_write_array(stream_t* stream, const void* array, size_t length);
